+++ client/src/App.jsx
@@ -16,10 +16,9 @@ import {
   IsAnon,
   GalleryPage,
   AddPicture,
-  TopicPage,
   CreateAPost,
   CreateATopic,
-  TopicDetails,
+  PostResults,
   SinglePost,
 } from './components';
 
@@ -33,8 +32,7 @@ function App() {
         <Route path="/word" element={<Words />} />
         <Route path="/gallery" element={<GalleryPage />} />
         <Route path="/gallery/add-picture" element={<AddPicture />} />
-        <Route path="/topics" element={<TopicPage />} />
-        <Route path="/topics/:id" element={<TopicDetails showAllPosts />} />
+        <Route path="/topics/:id" element={<PostResults showAllPosts />} />
         <Route path="/create-post" element={<CreateAPost />} />
         <Route path="/posts" element={<PostsPage />} />
         <Route path="/post/:id" element={<SinglePost />} />
