+++ client/src/components/Navbar/Navbar.jsx
@@ -52,7 +52,6 @@ export const Navbar = () => {
                   <OlinaLogo />
                 </Link>
                 <NavLink to="/gallery">Gallery</NavLink>
-                <NavLink to="/topics">Topics</NavLink>
                 <NavLink to="/posts">Posts</NavLink>
                 <NavLink to="/word">Dictionary</NavLink>
                 <NavLink to="/contact-me">Contact</NavLink>
