+++ client/src/components/Posts/PostResults.jsx
@@ -1,26 +1,25 @@
 import { useState, useContext, useRef, useLayoutEffect } from 'react';
-import { PostCard, CreateAPost } from '../../components';
+import { useQuery } from '@tanstack/react-query';
+import { PostCard, CreateAPost } from '..';
 import { AuthContext } from '../../context/auth.context';
 import { SearchContext } from '../../context/search.context';
-import { useQuery } from '@tanstack/react-query';
 import { postActions } from '../../services';
 
-export function TopicDetails({ id, selectedTopicId }) {
+export function PostResults({ selectedTopicId }) {
   const { isLoggedIn } = useContext(AuthContext);
   const { searchTerm } = useContext(SearchContext);
-  const [posts, setPosts] = useState([])
   const [isCreating, setIsCreating] = useState(false);
   const createPostRef = useRef(null);
 
   const { data: postsData = [], isFetching } = useQuery({
-    queryKey: ['posts', id],
+    queryKey: ['posts', selectedTopicId],
     queryFn: async () => {
-      const { posts: topicPosts } =
-        id === 'all'
+      const { posts, topic } =
+        selectedTopicId === 'all'
           ? await postActions.getPosts()
-          : await postActions.getPost(id);
+          : await postActions.getPostsByTopicId(selectedTopicId);
 
-      const filteredPosts = topicPosts.filter(
+      const filteredPosts = posts.filter(
         (post) =>
           post.body.toLowerCase().includes(searchTerm.toLowerCase()) ||
           post.title.toLowerCase().includes(searchTerm.toLowerCase()),
@@ -30,54 +29,10 @@ export function TopicDetails({ id, selectedTopicId }) {
         (a, b) => new Date(b.createdAt) - new Date(a.createdAt),
       );
 
-      return { posts: postSortedByDate };
+      return { posts: postSortedByDate, topic };
     },
-
-    // queryFn: async () => {
-    //   if (id === 'all') {
-    //     const { posts: topicPosts } = await postActions.getPosts();
-    //     const filteredPosts = topicPosts.filter(
-    //       (post) =>
-    //         post.body.toLowerCase().includes(searchTerm.toLowerCase()) ||
-    //         post.title.toLowerCase().includes(searchTerm.toLowerCase()),
-    //     );
-
-    //     const postSortedByDate = filteredPosts.sort(
-    //       (a, b) => new Date(b.createdAt) - new Date(a.createdAt),
-    //     );
-
-    //     return { posts: postSortedByDate };
-    //   } else {
-    //     const { topic, posts: topicPosts } = await postActions.getPost(id);
-
-    //     const filteredPosts = topicPosts.filter(
-    //       (post) =>
-    //         post.body.toLowerCase().includes(searchTerm.toLowerCase()) ||
-    //         post.title.toLowerCase().includes(searchTerm.toLowerCase()),
-    //     );
-
-    //     const postSortedByDate = filteredPosts.sort(
-    //       (a, b) => new Date(b.createdAt) - new Date(a.createdAt),
-    //     );
-
-    //     return { topic, posts: postSortedByDate };
-    //   }
-    // },
-
     staleTime: 1000 * 60 * 5,
   });
-
-  // useEffect(() => {
-  //   if (id !== 'all') {
-  //     api
-  //       .get(`/api/topic/details/${id}`)
-  //       .then((response) => {
-  //         setTopic(response?.data?.topic);
-  //         setPosts(response?.data?.topic?.posts);
-  //       })
-  //       .catch((err) => console.log(err));
-  //   }
-  // }, [id]);
 
   const handleCreatePost = () => {
     setIsCreating(!isCreating);
@@ -93,7 +48,7 @@ export function TopicDetails({ id, selectedTopicId }) {
     <div className="topic-details-container w-full px-4 bg-plum-400">
       <div className="flex justify-between items-center my-2">
         <h1 className="text-4xl font-bold text-gray-800 font-pop my-8 ml-24 w-full text-start">
-          {/* {!postsData.topic ? 'ALL POSTS' : 'TOPIC:  ' + postsData.topic.title} */}
+          {!postsData.topic ? 'ALL POSTS' : 'TOPIC:  ' + postsData.topic.title}
         </h1>
 
         {isLoggedIn && selectedTopicId !== 'all' && (
@@ -126,11 +81,7 @@ export function TopicDetails({ id, selectedTopicId }) {
       <div className="w-full">
         {isCreating && (
           <div ref={createPostRef}>
-            <CreateAPost
-              id={id}
-              posts={postsData.posts || []}
-              setPosts={setPosts}
-            />
+            <CreateAPost topicId={selectedTopicId} />
           </div>
         )}
       </div>
